backend/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
APScheduler==3.10.4
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
ecdsa==0.18.0
fastapi==0.104.1
greenlet==3.0.1
h11==0.14.0
idna==3.6
mysql-connector-python==8.2.0
protobuf==4.21.12
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
python-jose==3.3.0
pytz==2023.3.post1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.8.0
tzdata==2023.4
tzlocal==5.2
uvicorn==0.24.0.post1
